Add clear-all functionalitiy
